Move verdicts.py
